requirements.txt
Django>=5.0.6
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
Scrapy>=2.11.0
twisted>=23.10.0
pyOpenSSL>=24.0.0
lxml>=5.1.0
